DIAId / ChannelID are real structs now
Problem:
 DIAs currently have a size_t value that identifies the upstream data source.
 There is no knowledge wheter to call GetLocalBlocks() or GetNetworkBlocks() on
 DataManager. Somehow one must distinguish between those two address schemata.

Solution:
 ChainID is a struct that stores a type { LOCAL, NETWORK } and an identifier
 (size_t). One must provide a custom comparer to std::maps where ChainID is
 key. DiaID and ChannelId are typedefs to ChainID. Some components assert the
 type of the ChainID instances.

Future Work:
 DataManager provides a single method for iterators and emitters each.
 Depending on the type of the ChainID the call is redirected either to Channel-
 Multiplexer or the local BufferChainManager.

+++ c7a/net/channel.hpp
@@ -47,7 +47,7 @@ public:
 
     //! Creates a new channel instance
     Channel(NetDispatcher& dispatcher, ReleaseSocketCallback release_callback,
-            int id, int expected_streams, std::shared_ptr<data::BufferChain> target)
+            size_t id, int expected_streams, std::shared_ptr<data::BufferChain> target)
         : dispatcher_(dispatcher),
           release_(release_callback),
           id_(id),
@@ -82,7 +82,7 @@ public:
         return finished_streams_ == expected_streams_;
     }
 
-    int Id() {
+    size_t Id() {
         return id_;
     }
 
@@ -91,7 +91,7 @@ private:
     NetDispatcher& dispatcher_;
     ReleaseSocketCallback release_;
 
-    int id_;
+    size_t id_;
     int active_streams_;
     int expected_streams_;
     int finished_streams_;
